refactor(painters/dom): extract SDT helpers from renderer
Move SDT boundary computation, dataset application, snapshot
collection, and inline wrapper helpers out of renderer.ts into
dedicated modules under sdt/. Renames runs/sdt.ts to sdt/inline.ts
and re-exports remain stable. No behavior change.
